requirements.txt
Flask
PyYAML
selenium
webdriver-manager